backup_jobgenie/requirements.txt
Flask
openai
python-docx
docxtpl
python-dotenv
mammoth
